move: entity와 dto를 model 패키지 하위로 이동(#11)
