+++ src/main/java/com/dnd/namuiwiki/domain/survey/SurveyRepository.java
@@ -1,6 +1,6 @@
 package com.dnd.namuiwiki.domain.survey;
 
-import com.dnd.namuiwiki.domain.survey.entity.Survey;
+import com.dnd.namuiwiki.domain.survey.model.entity.Survey;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface SurveyRepository extends MongoRepository<Survey, String> {
