+++ src/main/java/com/dnd/namuiwiki/domain/survey/SurveyService.java
@@ -4,9 +4,9 @@ import com.dnd.namuiwiki.common.exception.ApplicationErrorException;
 import com.dnd.namuiwiki.common.exception.ApplicationErrorType;
 import com.dnd.namuiwiki.domain.option.OptionRepository;
 import com.dnd.namuiwiki.domain.question.QuestionRepository;
-import com.dnd.namuiwiki.domain.survey.dto.CreateSurveyRequest;
-import com.dnd.namuiwiki.domain.survey.dto.CreateSurveyResponse;
-import com.dnd.namuiwiki.domain.survey.entity.Survey;
+import com.dnd.namuiwiki.domain.survey.model.dto.CreateSurveyRequest;
+import com.dnd.namuiwiki.domain.survey.model.dto.CreateSurveyResponse;
+import com.dnd.namuiwiki.domain.survey.model.entity.Survey;
 import com.dnd.namuiwiki.domain.survey.model.SurveyAnswers;
 import com.dnd.namuiwiki.domain.survey.type.Period;
 import com.dnd.namuiwiki.domain.survey.type.Relation;
